Create Area of the Circle
Guido is a scientist. He was called to solve a problem in Ambrósio's factory. Mr Ambrósio wants to build a water treatment plant where the space allocation is the best possible. He gave him a paper which contains the radius of the places where he intends to build round tanks.

He expects to receive from Guido data that contains the area of each tank which whose radius were given.

However, Guido is full of work and asked for your help to write a program that calculates the area of each tank as fast as possible.

Assume that for this problem π = 3.14159. 

- Calculate the area.